+++ bundles/org.eclipse.swt/Eclipse SWT/cocoa/org/eclipse/swt/widgets/Tree.java
@@ -140,8 +140,8 @@ TreeItem _getItem (TreeItem parentItem, int index, boolean create) {
 	return item;
 }
 
-int accessibilityAttributeValue (int /*long*/ id, int /*long*/ sel, int /*long*/ arg0) {
-	
+int /*long*/ accessibilityAttributeValue (int /*long*/ id, int /*long*/ sel, int /*long*/ arg0) {
+
 	if (accessible != null) {
 		NSString attribute = new NSString(arg0);
 		id returnValue = accessible.internal_accessibilityAttributeValue(attribute, ACC.CHILDID_SELF);
@@ -873,7 +873,7 @@ void drawInteriorWithFrame_inView (int /*long*/ id, int /*long*/ sel, int /*long
 		GC gc = GC.cocoa_new (this, data);
 		gc.setFont (item.getFont (columnIndex));
 		if (isSelected) {
-			float /*double*/[] components = new float /*double*/[(int)/*64*/nsSelectionForeground.numberOfComponents ()];
+			float /*double*/ [] components = new float /*double*/[(int)/*64*/nsSelectionForeground.numberOfComponents ()];
 			nsSelectionForeground.getComponents (components);	
 			Color selectionForeground = Color.cocoa_new (display, components);
 			gc.setForeground (selectionForeground);
@@ -931,7 +931,7 @@ void drawInteriorWithFrame_inView (int /*long*/ id, int /*long*/ sel, int /*long
 	if (drawBackground && !drawSelection) {
 		NSGraphicsContext context = NSGraphicsContext.currentContext ();
 		context.saveGraphicsState ();
-		float[] colorRGB = background.handle;
+		float /*double*/ [] colorRGB = background.handle;
 		NSColor color = NSColor.colorWithDeviceRed (colorRGB[0], colorRGB[1], colorRGB[2], 1f);
 		color.setFill ();
 		NSBezierPath.fillRect (fullRect);
@@ -1107,7 +1107,7 @@ public int [] getColumnOrder () {
 	int [] order = new int [columnCount];
 	for (int i = 0; i < columnCount; i++) {
 		TreeColumn column = columns [i];
-		int index = ((NSOutlineView)view).columnWithIdentifier (column.nsColumn);
+		int index = (int)/*64*/((NSOutlineView)view).columnWithIdentifier (column.nsColumn);
 		if ((style & SWT.CHECK) != 0) index -= 1;
 		order [index] = i;
 	}
@@ -1588,7 +1588,7 @@ boolean isTrim (NSView view) {
  * right-clicks or control-clicks on an NSTableView or its subclasses. Fix is to select the 
  * clicked-on row ourselves.
  */
-int menuForEvent(int /*long*/ id, int /*long*/ sel, int /*long*/ theEvent) {
+int /*long*/ menuForEvent(int /*long*/ id, int /*long*/ sel, int /*long*/ theEvent) {
 	NSEvent event = new NSEvent(theEvent);
 	NSOutlineView tree = (NSOutlineView)view;
 	
@@ -1597,7 +1597,7 @@ int menuForEvent(int /*long*/ id, int /*long*/ sel, int /*long*/ theEvent) {
 	
 	// select the row that was clicked before showing the menu for the event
 	NSPoint mousePoint = view.convertPoint_fromView_(event.locationInWindow(), null);
-	int row = tree.rowAtPoint(mousePoint);
+	int /*long*/ row = tree.rowAtPoint(mousePoint);
 	
 	// figure out if the row that was just clicked on is currently selected
 	if (selectedRowIndexes.containsIndex(row) == false) {
@@ -2019,7 +2019,7 @@ boolean sendKeyEvent (NSEvent nsEvent, int type) {
 	return result;
 }
 
-void setBackground (float [] color) {
+void setBackground (float /*double*/ [] color) {
 	super.setBackground (color);
 	NSColor nsColor;
 	if (color == null) {
@@ -2084,7 +2084,7 @@ public void setColumnOrder (int [] order) {
 		for (int i=0; i<order.length; i++) {
 			int index = order [i];
 			TreeColumn column = columns[index];
-			int oldIndex = outlineView.columnWithIdentifier (column.nsColumn);
+			int /*long*/ oldIndex = outlineView.columnWithIdentifier (column.nsColumn);
 			int newIndex = i + check;
 			outlineView.moveColumn (oldIndex, newIndex);
 			newX [index] = (int)outlineView.rectOfColumn (newIndex).x;
@@ -2106,8 +2106,8 @@ public void setColumnOrder (int [] order) {
 void setFont(NSFont font) {
 	super.setFont (font);
 	if (!hooks (SWT.MeasureItem)) {
-		float ascent = font.ascender ();
-		float descent = -font.descender () + font.leading ();
+		float /*double*/ ascent = font.ascender ();
+		float /*double*/ descent = -font.descender () + font.leading ();
 		((NSOutlineView)view).setRowHeight ((int)Math.ceil (ascent + descent) + 1);
 	} else {
 		view.setNeedsDisplay (true);
